+++ frontend/portals/SearchFieldPortal/index.jsx
@@ -1,11 +1,22 @@
 import React, { useMemo } from 'react';
 import PropTypes from 'prop-types';
 import { useSelector } from 'react-redux';
-import { ConditionalWrapper } from '@shopgate/engage/components';
+import { ConditionalWrapper, ScrollHeader } from '@shopgate/engage/components';
 import { useRoute } from '@shopgate/engage/core';
+import { makeStyles } from '@shopgate/engage/styles';
 import { makeGetIsSearchBarVisible } from '../../selectors';
 import SearchField from '../../components/SearchField';
-import ScrollHeader from '../../components/ScrollHeader';
+import config from '../../config.json';
+
+const { hideOnScroll } = config;
+
+const useStyles = makeStyles()({
+  scrollHeader: {
+    top: 0,
+    zIndex: 100,
+    boxShadow: 'none',
+  },
+});
 
 /**
  * Renders SearchField component in app-bar.below.before portal
@@ -13,6 +24,7 @@ import ScrollHeader from '../../components/ScrollHeader';
  * @returns {JSX}
  */
 const SearchFieldPortal = ({ name }) => {
+  const { classes } = useStyles();
   const route = useRoute();
   const getIsSearchBarVisible = useMemo(makeGetIsSearchBarVisible, []);
   const isVisible = useSelector(state => getIsSearchBarVisible(state, {
@@ -28,7 +40,7 @@ const SearchFieldPortal = ({ name }) => {
     <ConditionalWrapper
       condition={name !== 'filter-bar.content.before'}
       wrapper={children => (
-        <ScrollHeader>
+        <ScrollHeader className={classes.scrollHeader} hideOnScroll={hideOnScroll}>
           {children}
         </ScrollHeader>
       )}